Setting scope variable name in OAuth1Client
Unfortunately not every web service using OAuth 1 uses the name 'scope' for the scope variable. For instance, Flickr uses 'perms' instead of 'scope'. Adding an option 'scopeName' to the JOAuth1Client class to change the name of the scope variable would allow the class to be used for those services that deviate from the standard in this matter.